requirements.txt
flask==3.0.3
flask-pymongo==2.3.0
pymongo==4.7.2
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
marshmallow==3.21.3
python-dotenv==1.0.1
pytest==8.2.2
pytest-flask==1.3.0
mongomock==4.1.2
